requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues


azure-functions
azure-storage-blob
flask==2.2.5
pandas==1.4.3
numpy==1.22.3
scikit-surprise==1.1.1
requests==2.28.1